+++ src/Core/NumArray.php
@@ -7,9 +7,25 @@ class NumArray
 {
     private $data;
 
+    private $shape;
+
     public function __construct(array $data)
     {
         $this->data = $data;
+    }
+
+    public function getShape(): array
+    {
+        if (is_null($this->shape)) {
+            $shape = [];
+            $currentDataDim = $this->data;
+            while (is_array($currentDataDim)) {
+                $shape[] = count($currentDataDim);
+                $currentDataDim = $currentDataDim[0];
+            }
+            $this->shape = $shape;
+        }
+        return $this->shape;
     }
 
     public function getData(): array